+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/dataanalysis/ValidationRuleExpressionDetails.java
@@ -1,4 +1,4 @@
-package org.hisp.dhis.analysis;
+package org.hisp.dhis.dataanalysis;
 
 /*
  * Copyright (c) 2004-2018, University of Oslo
